Re-organize CLI and add jmp-client sub-cli
