+++ jumpstarter/cli/util/__init__.py
@@ -1,4 +1,5 @@
 from .alias import AliasedGroup
+from .opt import opt_log_level
 from .table import make_table
 
-__all__ = ["AliasedGroup", "make_table"]
+__all__ = ["AliasedGroup", "make_table", "opt_log_level"]
