+++ jumpstarter/cli/util/alias.py
@@ -11,6 +11,11 @@ class AliasedGroup(click.Group):
         "move": ["mv"],
         "config": ["conf"],
         "delete": ["del"],
+        "shell": ["sh"],
+        "exporter": ["exporters", "e"],
+        "client": ["clients", "c"],
+        "lease": ["leases", "l"],
+        "version": ["ver", "v"]
     }
 
     def get_command(self, ctx: click.Context, cmd_name: str):
@@ -18,7 +23,8 @@ class AliasedGroup(click.Group):
         if rv is not None:
             return rv
         # Match if listed in the common aliases
-        matches = [x for x in self.list_commands(ctx) if self.common_aliases[x] and cmd_name in self.common_aliases[x]]
+        matches = [x for x in self.list_commands(ctx)
+                   if x in self.common_aliases and cmd_name in self.common_aliases[x]]
         if not matches:
             return None
         elif len(matches) == 1:
