+++ jumpstarter/cli/util/opt.py
@@ -1,0 +1,7 @@
+import click
+
+opt_log_level = click.option(
+       "-l", "--log-level", "log_level",
+       type=click.Choice(["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"]),
+       help="Log level"
+)
