+++ pyproject.toml
@@ -50,7 +50,8 @@ dev = [
 ]
 
 [project.scripts]
-jmp = "jumpstarter.cli.client:client"
+jmp = "jumpstarter.cli.jmp:jmp"
+jmp-client = "jumpstarter.cli.client:client"
 jmp-exporter = "jumpstarter.cli.exporter:exporter"
 j = "jumpstarter.cli.j:main"
